pipeline/requirements.txt
google-cloud-storage>=2.14.0
google-cloud-speech>=2.23.0
pydub>=0.25.1
google-generativeai>=0.3.0  # Optional; used for transcript cleaning and summarisation
ffmpeg-python>=0.2.0  # Optional; ensures ffmpeg is available in some environments